+++ academy-ui/src/app/api/[spaceId]/actions/spaces/update-tidbits-homepage/route.ts
@@ -1,18 +1,20 @@
-import { MutationUpdateTidbitsHomepageArgs } from '@/graphql/generated/generated-types';
-import { getSpaceById } from '@/app/api/helpers/space/getSpaceById';
+import { withErrorHandlingV1 } from '@/app/api/helpers/middlewares/withErrorHandling';
+import { getSpaceWithIntegrations } from '@/app/api/helpers/space';
 import { checkEditSpacePermission } from '@/app/api/helpers/space/checkEditSpacePermission';
-import { withErrorHandling } from '@/app/api/helpers/middlewares/withErrorHandling';
+import { getSpaceById } from '@/app/api/helpers/space/getSpaceById';
+import { MutationUpdateTidbitsHomepageArgs } from '@/graphql/generated/generated-types';
 import { prisma } from '@/prisma';
 import { SpaceTags } from '@/utils/api/fetchTags';
+import { Space, SpaceIntegration } from '@prisma/client';
 import { revalidateTag } from 'next/cache';
 import { NextRequest, NextResponse } from 'next/server';
 
-async function postHandler(req: NextRequest) {
+async function postHandler(req: NextRequest): Promise<NextResponse<{ space: Space & { spaceIntegrations: SpaceIntegration | null } }>> {
   const { spaceId, tidbitsHomepage } = (await req.json()) as MutationUpdateTidbitsHomepageArgs;
   const spaceById = await getSpaceById(spaceId);
   await checkEditSpacePermission(spaceById, req);
 
-  const space = await prisma.space.update({
+  await prisma.space.update({
     data: {
       tidbitsHomepage: {
         heading: tidbitsHomepage.heading,
@@ -25,8 +27,9 @@ async function postHandler(req: NextRequest) {
   });
 
   revalidateTag(SpaceTags.GET_SPACE.toString());
+  const space = await getSpaceWithIntegrations(spaceId);
 
   return NextResponse.json({ space }, { status: 200 });
 }
 
-export const POST = withErrorHandling(postHandler);
+export const POST = withErrorHandlingV1<{ space: Space & { spaceIntegrations: SpaceIntegration | null } }>(postHandler);
