+++ academy-ui/src/app/api/[spaceId]/queries/spaces/by-admin/route.ts
@@ -3,7 +3,7 @@ import { NextRequest, NextResponse } from 'next/server';
 import { prisma } from '@/prisma';
 import { Space } from '@prisma/client';
 
-export async function GET(req: NextRequest, { params }: { params: { spaceId: string } }): Promise<NextResponse<Space[] | { error: string }>> {
+export async function GET(req: NextRequest, { params }: { params: Promise<{ spaceId: string }> }): Promise<NextResponse<Space[] | { error: string }>> {
   const session = await getDecodedJwtFromContext(req);
   if (!session) return NextResponse.json({ error: 'No session found' }, { status: 401 });
