+++ academy-ui/src/app/clickable-demos/view/[...demoId]/page.tsx
@@ -1,54 +1,11 @@
-'use client';
+import { getSpaceServerSide } from '@/utils/space/getSpaceServerSide';
+import ViewClickableDemo from './ViewClickableDemo';
 
-import ClickableDemoModal from '@/components/clickableDemos/View/ClickableDemoModal';
-import withSpace from '@/contexts/withSpace';
-import { ClickableDemoWithSteps, SpaceWithIntegrationsFragment } from '@/graphql/generated/generated-types';
-import { SpaceTypes } from '@/types/space/SpaceDto';
-import FullPageLoader from '@dodao/web-core/components/core/loaders/FullPageLoading';
-import getBaseUrl from '@dodao/web-core/utils/api/getBaseURL';
-import axios from 'axios';
-import { useRouter } from 'next/navigation';
-import { useEffect, useState } from 'react';
-
-function ViewClickableDemo({ params, space }: { params: { demoId: string[] }; space: SpaceWithIntegrationsFragment }) {
+async function ViewClickableDemoPage(args: { params: Promise<{ demoId: string[] }> }) {
+  const params = await args.params;
   const demoId = params.demoId[0];
-  const [data, setData] = useState<ClickableDemoWithSteps>();
-  const [loading, setLoading] = useState<boolean>(false);
-  useEffect(() => {
-    async function fetchData() {
-      setLoading(true);
-      const response = await axios.get(`${getBaseUrl()}/api/${space.id}/clickable-demos/${demoId}`, {
-        params: {
-          spaceId: space.id,
-          demoId,
-        },
-      });
-
-      console.log('response.data: ', response.data);
-      setData(response.data);
-      setLoading(false);
-    }
-    fetchData();
-  }, [demoId, space.id]);
-
-  const router = useRouter();
-
-  if (loading) {
-    return <FullPageLoader />;
-  }
-
-  if (data) {
-    return (
-      <ClickableDemoModal
-        clickableDemoWithSteps={data}
-        space={space}
-        onClose={() => {
-          router.push(`${space.type === SpaceTypes.TidbitsSite ? '/' : '/clickable-demos'}`);
-        }}
-      />
-    );
-  }
-  return null;
+  const space = await getSpaceServerSide();
+  return <ViewClickableDemo space={space!} demoId={demoId} />;
 }
 
-export default withSpace(ViewClickableDemo);
+export default ViewClickableDemoPage;
