+++ base-ui/src/app/api/space/[id]/route.ts
@@ -1,7 +1,8 @@
 import { NextResponse } from 'next/server';
 import { prisma } from '@/prisma';
 
-export async function GET(req: Request, { params: { id } }: { params: { id: string } }) {
+export async function GET(req: Request, { params }: { params: Promise<{ id: string }> }) {
+  const { id } = await params;
   const space = await prisma.baseSpace.findFirstOrThrow({
     where: { id },
   });
